refacto(src): new architecture
