+++ .envrc
@@ -1,0 +1,3 @@
+source venv/bin/activate
+
+unset PS1
