+++ .github/ISSUE_TEMPLATE/bug_report.yml
@@ -1,0 +1,71 @@
+name: 🐞 Bug report
+description: If you have found a problem 🕵️
+title: "[Bug]: "
+labels: ["bug", "triage"]
+assignees:
+  - JBarmentlo
+body:
+  - type: markdown
+    attributes:
+      value: |
+        Thanks for taking the time to fill out this bug report!
+  - type: textarea
+    id: description
+    attributes:
+      label: Description
+      description: Describe your bug shortly.
+      placeholder: Tell us what you see!
+      value: "A bug happened!"
+    validations:
+      required: true
+  - type: textarea
+    id: what-happened
+    attributes:
+      label: What happened?
+      description: Describe your bug with as much details as possible.
+      placeholder: Tell us what you see!
+      value: "A bug happened!"
+    validations:
+      required: true
+  - type: textarea
+    id: expected-behaviour
+    attributes:
+      label: Expected behaviour
+      description: Describe what should happened with as much details as possible.
+      placeholder: Tell us what should happened!
+      value: "This should happened!"
+    validations:
+      required: true
+  - type: textarea
+    id: environment
+    attributes:
+      label: Environment
+      description: Describe your environment with as much details as possible.
+      placeholder: Tell us what is your environment!
+      value: "- OS and version\n- Branch that causes this bug"
+    validations:
+      required: true
+  - type: textarea
+    id: steps-to-reproduce
+    attributes:
+      label: Steps to reproduce
+      description: List the steps to reproduce with as much details as possible.
+      placeholder: Tell us how to reproduce this bug!
+      value: "- You have to do this!\n- Then this!"
+    validations:
+      required: true
+  - type: textarea
+    id: logs
+    attributes:
+      label: Relevant log output
+      description: Please copy and paste any relevant log output. This will be automatically formatted into code, so no need for backticks.
+      render: C++
+    validations:
+      required: false
+  - type: textarea
+    id: proposed-solution
+    attributes:
+      label: Proposed solution
+      description: If you have an idea of how to fix this issue, please write it down here, so we can begin discussing it.
+    validations:
+      required: false
