+++ .github/milestone_template.md
@@ -1,0 +1,65 @@
+## Overall
+
+### Objective
+
+[Explain here the objective of the milestone]
+
+### Timeline
+
+#### Duration
+
+- [ ] 1 week
+- [ ] 2 weeks
+- [ ] 3 weeks
+
+#### Date
+
+- Start date : [dd/mm/yyyy]
+- End date : [dd/mm/yyyy]
+
+### Checklist
+
+- [ ] Clear objective
+- [ ] Consistent objective
+- [ ] Achievable in the given time
+- [ ] Issues created, with the rights labels and linked to this milestone
+- [ ] Issues assigned
+
+## Monitoring
+
+### Week 1
+
+[Detail this week's report]
+
+### Week 2
+
+[Detail this week's report]\
+[If this milestone's duration is less than 2 weeks, please delete this section]
+
+### Week 3
+
+[Detail this week's report]\
+[If this milestone's duration is less than 3 weeks, please delete this section]
+
+## Final Report
+
+### Checklist
+
+- [ ] Objective fulfilled
+- [ ] README.md and other relevant documents (guide, ...) updated
+- [ ] Documentation updated
+- [ ] Pull requests merged
+- [ ] Issues closed
+- [ ] Release created
+- [ ] Tasks archived
+- [ ] Branches cleared
+
+### Questions
+
+**Was this objective fulfilled on time ?**
+
+[Answer this question]
+
+**If not, why ?**
+
+[Answer this question if needed, otherwise delete it]
