+++ .github/pull_request_template.md
@@ -1,0 +1,31 @@
+# Description
+
+Please provide a detailed description of what was done in this PR.
+Precise the issue that you are resolving.
+
+# Changes include
+
+- [ ] Bugfix (non-breaking change that solves an issue)
+- [ ] New feature (non-breaking change that adds functionality)
+- [ ] Breaking change (change that is not backwards-compatible and/or changes current functionality)
+
+# Breaking changes
+
+Please complete this section if any breaking changes have been made, otherwise delete it.
+
+# Checklist
+
+- [ ] I have assigned this PR to myself
+- [ ] I have added at least 1 reviewer
+- [ ] I have added the needed labels
+- [ ] I have linked this PR to an issue
+- [ ] I have linked this PR to a milestone
+- [ ] I have linked this PR to a project
+- [ ] I have tested this code
+- [ ] I have added / updated tests (unit / functionals / end-to-end / ...)
+- [ ] I have updated the README and other relevant documents (guides...)
+- [ ] I have added sufficient documentation both in code, as well as in the READMEs
+
+# Additional comments
+
+Please post additional comments in this section if you have them, otherwise delete it.
