+++ gym/Client.py
@@ -16,16 +16,15 @@ import time
 import colorama
 from colorama import Fore
 from matplotlib.pyplot import delaxes
-import defines
+from gym import defines
 
 HOST 			= "localhost"  # The server's hostname or IP address
 PORT 			= 13000        # The port used by the server
 PATH_TO_BOMBER 	= "/home/yup/Desktop/build/bomber.x86_64"
+PATH_TO_BOMBER = "simulator/build/bomber.x86_64"
 SLEEP_TIME		= 5.0 			# If the progRam must start the bomberbuddy it will wait this many seconds for bomberbuddy to start before attempting to connect
 
 dico = {0: "1", 1:"2",2:"B",3:"E",4:"W",5: "C", 6 : "r",7:"b",8:"s"}
-
-
 
 
 def itemtopos(item):
@@ -175,9 +174,3 @@ class  Client():
 
 		"""
 		return (self.board, self.player_states, self.winner)
-
-
-
-# c = Client()
-# c.connect()
-# c.request_type(-1)
